remove bad size_t

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: ngennaro <ngennaro@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 16:50:47 by ngennaro          #+#    #+#             */
-/*   Updated: 2022/11/22 13:31:25 by ngennaro         ###   ########lyon.fr   */
+/*   Updated: 2022/11/22 13:35:06 by ngennaro         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,8 +23,8 @@ static int	mul_overflow(long long int *dst, long long int a, long long int b)
 int	ft_atoi(const char *str)
 {
 	size_t			i;
-	size_t			j;
-	size_t			k;
+	int			j;
+	int				k;
 	long long int	res;
 
 	i = 0;
